phase 2 changes

+++ RealWorldApp/RealWorldApp/Models/OrderByUser.cs
@@ -20,10 +20,10 @@ namespace RealWorldApp.Models
         public DateTimeOffset OrderDate { get; set; }
         public Address ShipToAddress { get; set; }
         public string DeliveryMethod { get; set; }
-        public decimal ShippingPrice { get; set; }
+        public double ShippingPrice { get; set; }
         public IReadOnlyList<OrderItemDto> OrderItems { get; set; }
-        public decimal Subtotal { get; set; }
-        public decimal Total { get; set; }
+        public double Subtotal { get; set; }
+        public double Total { get; set; }
         public string Status { get; set; }
     }
 }
